Skip isoformat tests in Python 3.6
