+++ tests/entities.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from datetime import datetime
 from typing import (Collection,
                     Deque,
                     Dict,
@@ -15,7 +14,6 @@ from uuid import UUID
 from marshmallow import fields
 
 from dataclasses_json import DataClassJsonMixin, dataclass_json
-from dataclasses_json.mm import _IsoField
 
 A = TypeVar('A')
 
@@ -155,38 +153,10 @@ class DataClassJsonDecorator:
 
 @dataclass_json
 @dataclass
-class DataClassWithDatetime:
-    created_at: datetime
-
-
-@dataclass_json
-@dataclass
 class DataClassWithOverride:
     id: float = field(
         metadata={'dataclasses_json': {
             'mm_field': fields.Integer()
-        }})
-
-
-@dataclass_json
-@dataclass
-class DataClassWithIsoDatetime:
-    created_at: datetime = field(
-        metadata={'dataclasses_json': {
-            'encoder': datetime.isoformat,
-            'decoder': datetime.fromisoformat,
-            'mm_field': fields.DateTime(format='iso')
-        }})
-
-
-@dataclass_json
-@dataclass
-class DataClassWithCustomIsoDatetime:
-    created_at: datetime = field(
-        metadata={'dataclasses_json': {
-            'encoder': datetime.isoformat,
-            'decoder': datetime.fromisoformat,
-            'mm_field': _IsoField()
         }})
